requirements.txt
flask>=0.8
bunq-sdk>=0.12
